safari-534.53-branch/LayoutTests/canvas/philip/tests/2d.imageData.get.source.outside.html

<!DOCTYPE html>
<title>Canvas test: 2d.imageData.get.source.outside</title>
<script src="../tests.js"></script>
<link rel="stylesheet" href="../tests.css">
<body>
<p id="passtext">Pass</p>
<p id="failtext">Fail</p>
<p class="output">These images should be identical:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
_addTest(function(canvas, ctx) {

ctx.fillStyle = '#08f';
ctx.fillRect(0, 0, 100, 50);

var imgdata1 = ctx.getImageData(-10, 5, 1, 1);
_assertSame(imgdata1.data[0], 0, "imgdata1.data[\""+(0)+"\"]", "0");
_assertSame(imgdata1.data[1], 0, "imgdata1.data[\""+(1)+"\"]", "0");
_assertSame(imgdata1.data[2], 0, "imgdata1.data[\""+(2)+"\"]", "0");
_assertSame(imgdata1.data[3], 0, "imgdata1.data[\""+(3)+"\"]", "0");

var imgdata2 = ctx.getImageData(10, -5, 1, 1);
_assertSame(imgdata2.data[0], 0, "imgdata2.data[\""+(0)+"\"]", "0");
_assertSame(imgdata2.data[1], 0, "imgdata2.data[\""+(1)+"\"]", "0");
_assertSame(imgdata2.data[2], 0, "imgdata2.data[\""+(2)+"\"]", "0");
_assertSame(imgdata2.data[3], 0, "imgdata2.data[\""+(3)+"\"]", "0");

var imgdata3 = ctx.getImageData(200, 5, 1, 1);
_assertSame(imgdata3.data[0], 0, "imgdata3.data[\""+(0)+"\"]", "0");
_assertSame(imgdata3.data[1], 0, "imgdata3.data[\""+(1)+"\"]", "0");
_assertSame(imgdata3.data[2], 0, "imgdata3.data[\""+(2)+"\"]", "0");
_assertSame(imgdata3.data[3], 0, "imgdata3.data[\""+(3)+"\"]", "0");

var imgdata4 = ctx.getImageData(10, 60, 1, 1);
_assertSame(imgdata4.data[0], 0, "imgdata4.data[\""+(0)+"\"]", "0");
_assertSame(imgdata4.data[1], 0, "imgdata4.data[\""+(1)+"\"]", "0");
_assertSame(imgdata4.data[2], 0, "imgdata4.data[\""+(2)+"\"]", "0");
_assertSame(imgdata4.data[3], 0, "imgdata4.data[\""+(3)+"\"]", "0");


});
</script>
