LayoutTests/canvas/philip/tests/2d.imageData.get.source.outside-expected.txt

< [index] >
2d.imageData.get.source.outside
getImageData() returns transparent black outside the canvas
References: 2d.imageData.get.basic 2d.imageData.get.outside
Actual output:
Passed
